Created module CountValueInPostfix
