+++ Semester_1/Homework_5/Task_2/CountExpressionInPostfix.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <stdbool.h>
+
+bool countValueOfWholeExpression(char *inputExpression, double* resultingValue);
